🧹 refactor: keep transport diagnostics together
