+++ src/gh_llm/diagnostics.py
@@ -6,7 +6,6 @@ from typing import TYPE_CHECKING
 
 from gh_llm.environment import auth_status_command_text
 from gh_llm.invocation import display_command_with
-from gh_llm.transport_errors import looks_like_transport_error
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -20,6 +19,38 @@ _GRAPHQL_BACKED_COMMANDS = {
     ("gh", "pr", "view"),
     ("gh", "issue", "view"),
 }
+TRANSPORT_ERROR_PATTERNS = (
+    'post "https://api.github.com/graphql": eof',
+    "eof",
+    "timeout",
+    "i/o timeout",
+    "context deadline exceeded",
+    "client.timeout exceeded",
+    "request canceled",
+    "tls handshake timeout",
+    "remote error: tls",
+    "connection reset",
+    "connection reset by peer",
+    "connection refused",
+    "connection closed",
+    "connection aborted",
+    "broken pipe",
+    "temporary failure",
+    "temporarily unavailable",
+    "network is unreachable",
+    "server misbehaving",
+    "stream error",
+    "goaway",
+    "proxyconnect",
+    "http 500",
+    "http 502",
+    "http 503",
+    "http 504",
+    "500 internal server error",
+    "502 bad gateway",
+    "503 service unavailable",
+    "504 gateway timeout",
+)
 
 
 class GhCommandError(RuntimeError):
@@ -81,6 +112,11 @@ def format_command_error(error: GhCommandError) -> list[str]:
             lines.append(f"- {command}")
 
     return lines
+
+
+def looks_like_transport_error(message: str) -> bool:
+    lowered = message.lower()
+    return any(pattern in lowered for pattern in TRANSPORT_ERROR_PATTERNS)
 
 
 def _diagnose_command_error(error: GhCommandError) -> _Diagnosis:
